More embeddings, improved GUI functionality
I added a spectral and linear embedding (for the spectral embedding the java library Jama is needed - this has to be added to the build path). Furthermore, I also improved the functionality of the GUI a bit. Now, the menubar also contains a dropdown menu where the user can choose the layout of the graph and the GUI is resizable. The layout of the settings area still has to be adapted and the redrawing of the graph does not work in all cases